refactor : 일급 컬렉션 사용 & 불필요한 테스트 및 메쏘드 삭제
