+++ src/main/java/lotto/io/OutputView.java
@@ -5,6 +5,7 @@ import lotto.domain.LottoRank;
 import lotto.util.LottoRecorder;
 
 import java.util.List;
+import java.util.stream.Stream;
 
 public class OutputView {
     public static void printLotto(List<Lotto> lottoList) {
@@ -18,10 +19,11 @@ public class OutputView {
     public static void printResult(LottoRecorder lottoRecorder) {
         System.out.println("당첨 통계");
         System.out.println("---------");
-        System.out.println(String.format("%d개 일치(%d원)- %d개", LottoRank.FOURTH.getMatchingCount(), LottoRank.FOURTH.getPrice(), lottoRecorder.getLottoCount(LottoRank.FOURTH)));
-        System.out.println(String.format("%d개 일치(%d원)- %d개", LottoRank.THIRD.getMatchingCount(), LottoRank.THIRD.getPrice(), lottoRecorder.getLottoCount(LottoRank.THIRD)));
-        System.out.println(String.format("%d개 일치(%d원)- %d개", LottoRank.SECOND.getMatchingCount(), LottoRank.SECOND.getPrice(), lottoRecorder.getLottoCount(LottoRank.SECOND)));
-        System.out.println(String.format("%d개 일치(%d원)- %d개", LottoRank.JACKPOT.getMatchingCount(), LottoRank.JACKPOT.getPrice(), lottoRecorder.getLottoCount(LottoRank.JACKPOT)));
+        Stream.of(LottoRank.values())
+                .filter(lottoRank -> !lottoRank.equals(LottoRank.REST))
+                .forEach(lottoRank -> 
+                    System.out.println(String.format("%d개 일치(%d원)- %d개", lottoRank.getMatchingCount(), lottoRank.getPrice(), lottoRecorder.getLottoCount(lottoRank)))
+                );
         System.out.println(String.format("총 수익률은 %d입니다.",lottoRecorder.getProfitRatio()));
     }
 }
