+++ src/main/java/lotto/util/LottoRecorder.java
@@ -6,17 +6,40 @@ import lotto.domain.LottoResult;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class LottoRecorder {
 
+    private Lotto jackpot;
     private LottoResult lottoResult;
 
     public LottoRecorder(Lotto jackpot, List<Lotto> lottoList) {
-        this.lottoResult = new LottoResult(jackpot, lottoList);
+        Objects.requireNonNull(jackpot);
+        this.jackpot = jackpot;
+        this.lottoResult = new LottoResult();
+        recordResult(lottoList);
+    }
+
+    private void recordResult(List<Lotto> lottoList) {
+        Stream.of(LottoRank.values()).forEach(lottoRank ->
+                lottoResult.put(lottoRank, matchCount(lottoRank, lottoList))
+        );
+    }
+
+    private List<Lotto> matchCount(LottoRank lottoRank, List<Lotto> lottoList) {
+        return lottoList.stream()
+                .filter(lotto ->
+                        lottoRank.equals(LottoRank.REST) ?
+                                matchCount(lotto) < 3 :
+                                lottoRank.getMatchingCount() == matchCount(lotto)
+                )
+                .collect(Collectors.toList());
     }
 
     public int matchCount(Lotto lotto) {
-        return lottoResult.matchCount(lotto);
+        return jackpot.containCount(lotto);
     }
 
     public int getLottoCount(LottoRank lottoRank) {
